refactor: change component folder structure
