+++ app/(main)/layout.tsx
@@ -1,5 +1,5 @@
-import Footer from './_components/Footer'
-import Header from './_components/Header'
+import Footer from '@/components/common/Footer'
+import Header from '@/components/common/Header'
 
 export default function MainLayout({
   children
