+++ app/(main)/study/page.tsx
@@ -1,4 +1,4 @@
-import SectionBanner from '@/components/SectionBanner'
+import SectionBanner from '@/components/common/SectionBanner'
 
 export default function Study() {
   return (
